Move torch-dependent code out of util.py into torch_util.py
